+++ app/Http/Controllers/Authors/AuthorController.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Authors;
 
+use App\Models\Author;
+use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreAuthorRequest;
 use App\Http\Requests\UpdateAuthorRequest;
-use App\Models\Author;
 
 class AuthorController extends Controller
 {
